bau: Provide more help diagnosing key issues
A common issue we see with people using Verify Service Provider is
misconfiguring the keys so the private keys in its configuration don't
match the certificates Verify is expecting to see. This can sometimes
result in some quite tricky to diagnose issues.

Longer term it would be nice for the self-service application for
uploading certificates to also print the checksums of the keys it's
expecting to see to make it as easy as possible to spot when the keys
don't match up. In the absence of such and application we'll have to
make do with documenting the openssl invocations required.